libmisc: list: update helpers
Add a free_list helper so that we can duplicate list to callers and
require them to do the freeing for us, as well as a comma_from_list
helper for the /etc/sub{uid,gid} writing functions. In addition, fix up
the duplication code to just ignore NULL lists rather than aborting (for
no good reason).

As an aside, we really should switch to sharing the list code from
somewhere else rather than maintaining a separate version within
shadow-utils...

Signed-off-by: Aleksa Sarai <asarai@suse.de>